requirements.txt
flask==1.1.1
pandas==0.25.0
gensim==3.8.0
fasthan==1.3
opencc-python-reimplemented==0.1.6
tqdm==4.46.1
jieba==0.39
numpy==1.17.0
torch==1.5.0+cu101